Align master with gh-2567-rewrite-docs
