+++ samples/boot/helloworld/spring-security-samples-boot-helloworld.gradle
@@ -9,5 +9,4 @@ dependencies {
 
 	testCompile project(':spring-security-test')
 	testCompile 'org.springframework.boot:spring-boot-starter-test'
-	testCompile seleniumDependencies
 }
