+++ samples/boot/helloworld/src/main/resources/application.yml
@@ -1,4 +1,12 @@
+server:
+  port: 8080
+
+logging:
+  level:
+    root: WARN
+    org.springframework.web: INFO
+    org.springframework.security: INFO
+
 spring:
-  security:
-    user:
-      password: password
+  thymeleaf:
+    cache: false
