+++ samples/javaconfig/hellomvc/src/test/java/org/springframework/security/samples/pages/WebSecurityConfigTests.java
@@ -1,0 +1,89 @@
+/*
+ * Copyright 2002-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.security.samples.pages;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestBuilders.formLogin;
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestBuilders.logout;
+import static org.springframework.security.test.web.servlet.response.SecurityMockMvcResultMatchers.unauthenticated;
+import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+/**
+ * @author Rob Winch
+ *
+ */
+@RunWith(SpringJUnit4ClassRunner.class)
+@ContextConfiguration(classes = { MvcConfiguration.class, WebSecurityConfig.class })
+@WebAppConfiguration
+public class WebSecurityConfigTests {
+	private MockMvc mvc;
+
+	@Autowired
+	private WebApplicationContext context;
+
+
+	@Before
+	public void setup() {
+		mvc = MockMvcBuilders.webAppContextSetup(context)
+				.apply(springSecurity())
+				.defaultRequest(get("/").accept(MediaType.TEXT_HTML)).build();
+	}
+
+	@Test
+	public void requestProtectedResourceRequiresAuthentication() throws Exception {
+		mvc.perform(get("/")).andExpect(redirectedUrl("http://localhost/login"));
+	}
+
+	@Test
+	public void loginSuccess() throws Exception {
+		mvc.perform(formLogin()).andExpect(redirectedUrl("/"));
+	}
+
+	@Test
+	public void loginFailure() throws Exception {
+		mvc.perform(formLogin().password("invalid")).andExpect(
+				redirectedUrl("/login?error"));
+	}
+
+	@Test
+	@WithMockUser
+	public void requestProtectedResourceWithUser() throws Exception {
+		mvc.perform(get("/")).andExpect(status().isOk());
+	}
+
+	@Test
+	@WithMockUser
+	public void logoutSuccess() throws Exception {
+		mvc.perform(logout()).andExpect(redirectedUrl("/login?logout"))
+				.andExpect(unauthenticated());
+	}
+}
